# Improve performance of shader lighting code in Forward renderers.