… servers/rendering/renderer_rd/shaders/scene_forward_lights_inc.glsl