- Skip sampling shadows if attenuation is very small.
- Skip computing diffuse and specular light if attenuation and shadow are very small.

Using the TPS demo from this [PR](https://github.com/godotengine/godot/pull/90284) with the Mobile renderer.

Approximate GPU times.

Ryzen 7950 Integrated Graphics
`master`: ~32.5 ms
`light-compute-attenuation-skip`: ~27.0 ms

Mali G-78 (Samsung S21)
`master`: ~27.4 ms
`light-compute-attenuation-skip`: ~24.6 ms

Mali G-715 (Google Pixel 8)
`master`: ~19.4 ms
`light-compute-attenuation-skip`: ~17.3 ms

Behavior should be identical in theory although verification is welcome. Performance improvements are also expected in Forward+ but will be less significant as there's light clustering involved.

Opening as draft for now as I intend to do more research into improving some of the other features by reorganizing the code further.

---

Contributed by [W4 Games](https://w4games.com/). 🍀