@clayjohn 2024-12-04
Finally tested on an older device (Pixel 4 Adreno 640). This PR roughly reduces the frame time from 285 ms to about 260 ms which lines up with the other mobile devices you tested on.